Partial support for blocks higher than 256 id

+++ src/com/lishid/orebfuscator/obfuscation/BlockUpdate.java
@@ -91,7 +91,7 @@ public class BlockUpdate {
 
 	@SuppressWarnings("deprecation")
 	public static void AddBlockCheck(List<Block> allBlocks, Block block) {
-		if ((OrebfuscatorConfig.isObfuscated((byte) block.getTypeId(), block.getWorld().getEnvironment() == Environment.NETHER) || OrebfuscatorConfig.isDarknessObfuscated((byte) block.getTypeId()))) {
+		if ((OrebfuscatorConfig.isObfuscated(block.getTypeId(), block.getWorld().getEnvironment() == Environment.NETHER) || OrebfuscatorConfig.isDarknessObfuscated((byte) block.getTypeId()))) {
 			allBlocks.add(block);
 		}
 	}
